+++ src/core/index.ts
@@ -5,3 +5,7 @@ export * from "./types";
 export * from "./placeholders";
 export * from "./helpers";
 export * from "./constants";
+export * from "./conditions";
+export * from "./formatting";
+
+export { iterateControls } from "./util";